% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNHC.R
\name{calNHC}
\alias{calNHC}
\title{Subtracts Housing Expenditures from Total Expenditures.
Saves this in $NHC column.}
\usage{
calNHC(arg = TRUE)
}
\arguments{
\item{arg}{Sentence describing arg.
#' @keywords rename, parse}
}
\description{
This function subtracts.
}
\examples{
calcNHC(TRUE)
calcNHC(FALSE)
}

