% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusFunction1.R, R/censusFunction2.R
\name{censusFunction1}
\alias{censusFunction1}
\title{A Census Function}
\usage{
censusFunction1(year = TRUE)

censusFunction1(year = TRUE)
}
\arguments{
\item{year}{The year which you will perform the function on.}

\item{year}{The year that data will be returned for.}
}
\description{
This function allows you to do something.

This function allows you to do something.
}
\examples{
censusFunction1()
censusFunction1(1976)
censusFunction1(1977)
censusFunction1(1978)
censusFunction2()
}
\keyword{household}
\keyword{individual,}

