% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusGraph.R
\name{censusGraph}
\alias{censusGraph}
\title{Plot visual graphs of data.}
\usage{
censusGraph(plot)
}
\arguments{
\item{plot}{Name of the plot to show. Plot names are in censusFunctions::createGraphs}
}
\description{
This function displays processed data visually.
}
\examples{
censusGraph("allIncome")
censusGraph("allConsumption")
censusGraph("car")
}
\keyword{display,}
\keyword{graph,}
\keyword{plot,}
\keyword{viewer}

