% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chngDir.R
\name{chngDir}
\alias{chngDir}
\title{Helper function to change directories.}
\usage{
chngDir(arg)
}
\arguments{
\item{arg}{Paper or package?}
}
\description{
This function moves between 2 directories.
}
\examples{
chngDir(package)
chngDir(febThesis)
chngDir(paper)
}
\keyword{dir,}
\keyword{directory}
