% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineFrames.R
\name{combineFrames}
\alias{combineFrames}
\title{combine from raw data.}
\usage{
combineFrames(arg = TRUE)
}
\arguments{
\item{arg}{Sentence describing arg.
#' @keywords combine, dataframe, year}
}
\description{
This function combines multiple years to a dataframe.
}
\examples{
combineFrames(TRUE)
combineFrames(FALSE)
combineFrames()
}
