% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineHousingPrices.R
\name{combineHousingFunctions}
\alias{combineHousingFunctions}
\title{combine housing prices from multiple excel files}
\usage{
combineHousingFunctions(arg1)
}
\arguments{
\item{arg1}{What am I creating? Defaults to saved.}
}
\description{
This function imports Israeli census data.
}
\examples{
combineHousingPrices("saved")
combineHousingPrices("excel")
combineHousingPrices("raw")
combineHousingPrices("writeout")
}
\keyword{aggregate,}
\keyword{data}
\keyword{process,}
