% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineHousingPrices.R, R/create.R
\name{create}
\alias{create}
\alias{create}
\title{combine housing prices from multiple excel files}
\usage{
create(arg1)

create(arg1)
}
\arguments{
\item{arg1}{What am I creating? Defaults to saved.}

\item{arg1}{What am I creating? Defaults to matrices.}
}
\description{
This function imports Israeli census data.

This function imports Israeli census data.
}
\examples{
combineHousingPrices("saved")
combineHousingPrices("excel")
combineHousingPrices("raw")
combineHousingPrices("writeout")
create('matrices')
create('means')
}
\keyword{aggregate,}
\keyword{data}
\keyword{process,}
