% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraphs.R
\name{createGraphs}
\alias{createGraphs}
\title{Create ggplot2 plots, but don't print to screen/page.}
\usage{
createGraphs(arg1 = 1)
}
\arguments{
\item{arg1}{defaults to TRUE}
}
\description{
This function aggregates graphing code; returns plots as a list.
}
\examples{
createGraphs(1)
createGraphs(0)
}
\keyword{graphs}
\keyword{plots,}

