% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlplethr.R
\name{dlplethr}
\alias{dlplethr}
\title{dlplethr v.0.3}
\usage{
dlplethr(source = "saved")
}
\description{
This function modifies a map of Israel, adding the Golan.
}
\examples{
dlplethr("raw")
dlplethr("saved")
}
\keyword{GIS}
\keyword{choropleth,}
\keyword{choroplethr,}
\keyword{import,}
\keyword{map,}
\keyword{region,}
