% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloadHousingPrices.R
\name{downloadHousingPrices}
\alias{downloadHousingPrices}
\title{Downloads quarterly housing prices for the years
2006-2016 to the current working directory as 5 excel files}
\usage{
downloadHousingPrices(key = 3)
}
\arguments{
\item{year}{defaults to true}
}
\description{
Downloads quarterly housing prices for the years
2006-2016 to the current working directory as 5 excel files
}
\examples{
downloadHousingPrices(1)        #downloads Excel files from CBS servers
downloadHousingPrices(FALSE)    #displays a message
downloadHousingPrices(3)       #downloads a test file
downloadHousingPrices()         #downloads a test file
}
\keyword{CBS,}
\keyword{cpi,}
\keyword{dataset}
\keyword{download,}
\keyword{fetch,}
\keyword{housing,}
\keyword{price,}

