% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econClusters.R
\name{econClusters}
\alias{econClusters}
\title{subsets by econ cluster v: 2.0}
\usage{
econClusters(arg = TRUE)
}
\arguments{
\item{arg}{nil}
}
\description{
subsets by econ cluster v: 2.0
}
\examples{
econCluster()
econCluster(TRUE)
}
\keyword{econ}
\keyword{level,}
\keyword{subset}
