% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Import data to workspace v.0.5 correct subsetting}
\usage{
importData(source)
}
\arguments{
\item{source}{Source for data, "raw" imports from .por files. "saved" files were previously parsed in R.}
}
\description{
This function imports Israeli census data from spss .por files distributed by the CBS.
}
\examples{
importData("raw")
importData("saved")
importData("writeOut")
importData("family")
}
\keyword{import,}
\keyword{load}
\keyword{por,}
\keyword{raw,}
\keyword{saved,}
\keyword{spss,}

