% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_if_needed.R
\name{install_if_needed}
\alias{install_if_needed}
\title{A wrapper for install.}
\usage{
install_if_needed(pkgs, lib = NULL, repos = getOption("repos"),
  contriburl = contrib.url(repos, type), method, available = NULL,
  destdir = NULL, dependencies = NA, type = getOption("pkgType"),
  configure.args = getOption("configure.args"),
  configure.vars = getOption("configure.vars"), clean = FALSE,
  Ncpus = getOption("Ncpus", 1L), verbose = getOption("verbose"),
  libs_only = FALSE, INSTALL_opts, quiet = FALSE, keep_outputs = FALSE,
  check_version = TRUE, ...)
}
\description{
This function allows you to install packages.
}
\examples{
install_if_needed('dplyr')

}
\keyword{install}
