% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installCensusFunctions.R
\name{installCensusFunctions}
\alias{installCensusFunctions}
\title{Helper function to change directories. v.1.3}
\usage{
installCensusFunctions(location)
}
\arguments{
\item{location}{Either local or github.}
}
\description{
This function installs the local or github version of Israel Census helper functions.
}
\examples{
installCensusFunctions('local')
installCensusFunctions('github')
}
\seealso{
Other setup: \code{\link{install_if_needed}},
  \code{\link{makeFamiliesList}},
  \code{\link{maxReproducible}}, \code{\link{preSetUp}},
  \code{\link{setUp}}
}
\keyword{dir,}
\keyword{directory}
