% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeilframe.R
\name{makeilframe}
\alias{makeilframe}
\title{makeilframe v.0.4}
\usage{
makeilframe(x = "saved")
}
\description{
This function is a wrapper around dlplethr and modplethr.
}
\examples{
makeilframe("raw")
makeilframe("saved")
makeilframe("alt")
}
\keyword{GIS}
\keyword{choropleth,}
\keyword{choroplethr,}
\keyword{import,}
\keyword{map,}
\keyword{region,}
