% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxReproducible.R
\name{maxReproducible}
\alias{maxReproducible}
\title{Work with saved data or from raw census downloads?  v: 1.2
Defaults to saved. 1 for raw. 2 for saved. 3 to import more saved.  4. to write out what was made by two.}
\usage{
maxReproducible(level = 2)
}
\description{
Work with saved data or from raw census downloads?  v: 1.2
Defaults to saved. 1 for raw. 2 for saved. 3 to import more saved.  4. to write out what was made by two.
}
\examples{
maxReproducible(1)
maxReproducible(2)
maxReproducible(3)
maxReproducible(4)
maxReproducible()
}
\keyword{import,}
\keyword{reproducibility}
\keyword{saved,}
