% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFrames.R
\name{mergeFrames}
\alias{mergeFrames}
\title{Merge all years (in familiesList) into a single dataframe to facilitate regression.}
\usage{
mergeFrames(arg = 3)
}
\arguments{
\item{arg}{Defaults to 1. 2 is not implemented.}
}
\description{
This function combines multiple years to a dataframe. v: 3.2
}
\examples{
mergeFrames("merge")
mergeFrames("writeOut")
mergeFrames()
mergeFrames("saved")
}
\seealso{
Other combine: \code{\link{combineFrames}}
}
\keyword{combine,}
\keyword{dataframe,}
\keyword{merge,}
\keyword{regression,}
\keyword{year}
