% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modplethr.R
\name{modplethr}
\alias{modplethr}
\title{modplethr v.0.3}
\usage{
modplethr(args = "saved")
}
\description{
This function modifies a map of Israel, adding the Golan.
}
\examples{
modplethr("raw")
modplethr("saved")
modplethr("write")
}
\keyword{GIS}
\keyword{choropleth,}
\keyword{choroplethr,}
\keyword{import,}
\keyword{map,}
\keyword{region,}
