% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverage.R
\name{movingAverage}
\alias{movingAverage}
\title{Add Moving Average Column to YYYY. v. 0.3}
\usage{
movingAverage(arg = TRUE)
}
\arguments{
\item{arg}{The year that data will be returned for.}
}
\description{
This function creates a moving average of homeP.
}
\examples{
movingAverage()
requires:
1: censusFunctions::importData("versionF")
2: load("../savedData/homeprices.Rda")
3: source("../includes/regionalPrices.R")
movingAverage("load")
movingAverage("bar")
}
\seealso{
Other mutate: \code{\link{regionalSizePrices}}
}
\keyword{average,}
\keyword{houseP}
