% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownership.R
\name{ownership}
\alias{ownership}
\title{Divide into owners and renters.
Is called by: weighHouseholds(), takes expYYYYs
Creates dataframes based on payment of rent.}
\usage{
ownership(class, year)
}
\arguments{
\item{class}{owners or renters, defaults to owners.}

\item{year}{Numeric, from 2004 to 2014}
}
\description{
Divide into owners and renters.
Is called by: weighHouseholds(), takes expYYYYs
Creates dataframes based on payment of rent.
}
\examples{
ownership("renters", 2012)
ownership("owners",  2004)
}
\keyword{owner,}
\keyword{renter,}
\keyword{subset}

