% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownershipByCity.R
\name{ownershipByCity}
\alias{ownershipByCity}
\title{split ownersYYYY and rentersYYYY by city  v.2.3
This function creates subsets of renters and owners by city
call with arg (2) to split by median income}
\usage{
ownershipByCity(arg = 1)
}
\arguments{
\item{arg}{Defaults to 1}
}
\description{
split ownersYYYY and rentersYYYY by city  v.2.3
This function creates subsets of renters and owners by city
call with arg (2) to split by median income
}
\examples{
ownershipByCity(1)
ownershipByCity(2)
ownershipByCity()
}
\keyword{owners,}
\keyword{renters,}
\keyword{subset}
