% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownersIncome.R
\name{ownersIncome}
\alias{ownersIncome}
\title{Divide into owners and renters.}
\usage{
ownersIncome(arg = TRUE)
}
\arguments{
\item{arg}{Placeholder.}
}
\description{
Creates dataframes based on payment of rent..
}
\examples{
ownersIncome()
}
\keyword{owner,}
\keyword{renter,}
\keyword{subset}

