% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preSetUp.R
\name{preSetUp}
\alias{preSetUp}
\title{initial setup code
This function wraps setup code}
\usage{
preSetUp(arg = TRUE)
}
\arguments{
\item{arg}{Defaults to true.}
}
\description{
initial setup code
This function wraps setup code
}
\examples{
preSetUp(TRUE)
preSetUp()
}
\keyword{setup}
