% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameColumns.R
\name{renameColumns}
\alias{renameColumns}
\title{rename columns from raw data.}
\usage{
renameColumns(arg = TRUE)
}
\arguments{
\item{arg}{Sentence describing arg.
#' @keywords rename, parse}
}
\description{
This function imports Israeli census data.
}
\examples{
renameColumns(TRUE)
renameColumns(FALSE)
renameColumns()
}
