% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentersIncome.R
\name{rentersIncome}
\alias{rentersIncome}
\title{Divide into owners and renters. v. .79 - doesn't run}
\usage{
rentersIncome(arg = TRUE)
}
\arguments{
\item{arg}{Placeholder.}
}
\description{
Creates dataframes based on payment of rent..
}
\examples{
rentersIncome()
}
\keyword{owner,}
\keyword{renter,}
\keyword{subset}
