% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savedGraphs.R
\name{savedGraphs}
\alias{savedGraphs}
\title{Graphs I've made before  v0.1.3}
\usage{
savedGraphs(class = "renters", x = "income", y = "consumption",
  year = 2004)
}
\arguments{
\item{class}{Owner or renter? Defaults to owner.}

\item{x}{What goes on the x-axis?}

\item{y}{What goes on the y-axis?}

\item{year}{What year do you want to see?}
}
\description{
This function plots with ggplot2
}
\examples{
savedGraphs('owners', 'income', 'consumption', 2005)
savedGraphs('renters','consumption',   'rent', 2013)
savedGraphs('renters','income',        'rent', 2010)
}
\keyword{graph,}
\keyword{plot}
