% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{setUp}
\alias{setUp}
\title{Secondary setup code, replaces setupV7 v. 1.5
This function wraps setup code, arg = 6 is current version.}
\usage{
setUp(arg = 4)
}
\arguments{
\item{arg}{Defaults to 4.}
}
\description{
Secondary setup code, replaces setupV7 v. 1.5
This function wraps setup code, arg = 6 is current version.
}
\examples{
setUp()
setUp(1)
setUp(2)
setUp(3)
setUp(4)
setUp(5)
setUp(6)
}
\seealso{
Other setup: \code{\link{installCensusFunctions}},
  \code{\link{install_if_needed}},
  \code{\link{makeFamiliesList}},
  \code{\link{maxReproducible}}, \code{\link{preSetUp}}
}
\keyword{setup}
