% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\title{A template for function creation}
\usage{
skeleton(arg = 4)
}
\arguments{
\item{arg}{Defaults to 4.}
}
\description{
A template for function creation
}
\examples{
skeleton(1)
skeleton(2)
}
\keyword{setup}
