% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceRmd.R
\name{sourceRmd}
\alias{sourceRmd}
\title{Source another chapter without switching buffers v.0.1}
\usage{
sourceRmd(file, local = FALSE, ...)
}
\description{
Time saver for interactive use.
}
\examples{
sourceRmd()
}
\keyword{eval}
\keyword{load,}
\keyword{source,}
