% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockData.R
\name{stockData}
\alias{stockData}
\title{Import stock data from exchange v. 2.3}
\usage{
stockData(arg)
}
\arguments{
\item{arg}{import- get it, add- add it as a column to the data frame, write- write the new dataframe}
}
\description{
This function imports mid-year prices from the tlv-125 index.
}
\examples{
stockData('import')
stockData('writeOutStock')
stockData('saved')
stockData('createSaved')
stockData('add')
stockData('lagged')
}
\seealso{
Other manipulate: \code{\link{makeRegions}},
  \code{\link{subsetCities}}
}
\keyword{import,}
\keyword{non-housing}
\keyword{stocks,}
\keyword{tlv-125,}
\keyword{wealth,}
\keyword{write}
