% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByOwnership.R
\name{subsetByOwnership}
\alias{subsetByOwnership}
\title{Creates dataframes expYYYYowners and expYYYYrenters,
calls weighHouseholds() ;   v. 3.8}
\usage{
subsetByOwnership(arg = TRUE)
}
\arguments{
\item{arg}{Defaults to true.}
}
\description{
Creates dataframes expYYYYowners and expYYYYrenters,
calls weighHouseholds() ;   v. 3.8
}
\examples{
subsetByOwnership('listKeyMoney')
subsetByOwnership()

## this should be called after loadData("versionF")
subsetByOwnership('versionF')

## this was the previous version
subsetByOwnership('versionD')
subsetByOwnership('better')

subsetByOwnership('listOwners')
subsetByOwnership('listNotKey')
subsetByOwnership('listHouseholds')
subsetByOwnership('listRenters')
subsetByOwnership('listKeyMoney')
subsetByOwnership(1)
subsetByOwnership(FALSE)
}
\keyword{owners,}
\keyword{renters,}
\keyword{subset}
