% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyFrames.R
\name{tidyFrames}
\alias{tidyFrames}
\title{Applies Weights to households.}
\usage{
tidyFrames(metric)
}
\arguments{
\item{metric}{Takes values of income, consumption or NHC.}
}
\description{
This function weighs Israeli census data. Returns 1 of 6 options mat[owner/renter]Weighted[Income/Consumption/NHC].
}
\examples{
tidyFrame('income')
tidyFrame('consumption')
tidyFrame('NHC')
}
\keyword{aggregate,}
\keyword{data}
\keyword{process,}
