% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNHC.R
\name{calcNHC}
\alias{calcNHC}
\title{Subtracts Housing Expenditures from Total Expenditures. v. 0.4
Saves this in $NHC column.}
\usage{
calcNHC(arg = "exp")
}
\arguments{
\item{arg}{Sentence describing arg.}
}
\description{
This function subtracts.
}
\examples{
calcNHC()
calcNHC("versionG")
calcNHC("exp")
}
\seealso{
Other manipulate: \code{\link{makeRegions}},
  \code{\link{stockData}}, \code{\link{subsetCities}}
}
\keyword{parse}
\keyword{rename,}
