% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusFunction1.R
\name{censusFunction1}
\alias{censusFunction1}
\title{A Census Function}
\usage{
censusFunction1(year = TRUE)
}
\description{
This function allows you to do something.
}
\examples{
censusFunction1()
}
\keyword{household}
\keyword{individual,}

