% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineFrames.R
\name{combineFrames}
\alias{combineFrames}
\title{combine all years into a single dataframe to facilitate regression.}
\usage{
combineFrames(arg = TRUE)
}
\arguments{
\item{arg}{Sentence describing arg.
#' @keywords combine, dataframe, year}
}
\description{
This function combines multiple years to a dataframe. v: 2.0
}
\examples{
combineFrames(TRUE)
combineFrames(FALSE)
combineFrames()
}
