% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Gather statistics from multiple years and create a matrix}
\usage{
create(arg1)
}
\arguments{
\item{arg1}{What am I creating? Defaults to matrices.}
}
\description{
This function imports Israeli census data.
}
\examples{
create("matrices")
create("means")
}
\keyword{aggregate,}
\keyword{data}
\keyword{process,}

