% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraphs.R
\name{createGraphs}
\alias{createGraphs}
\title{Create ggplot2 plots, but don't print to screen/page.}
\usage{
createGraphs(metric = "income")
}
\arguments{
\item{metric}{income/consumption/NHC ;; weighted}
}
\description{
This function aggregates graphing code; returns plots as a list.
}
\examples{
createGraphs("income")
createGraphs("NHC")
createGraphs("consumption")
createGraphs()
}
\seealso{
Other plot: \code{\link{savedGraphs}}
}
\keyword{graphs}
\keyword{plots,}
