% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowloadHousingPrices.R
\name{downloadHousingPrices}
\alias{downloadHousingPrices}
\title{Downloads quarterly housing prices for the years
2006-2016 to the current working directory as 5 excel files}
\usage{
downloadHousingPrices(year = TRUE)
}
\arguments{
\item{year}{defaults to true}
}
\description{
Downloads quarterly housing prices for the years
2006-2016 to the current working directory as 5 excel files
}
\examples{
downloadHousingPrices()
}
\keyword{CBS,}
\keyword{cpi,}
\keyword{dataset}
\keyword{download,}
\keyword{fetch,}
\keyword{housing,}
\keyword{price,}

