% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econClusters.R
\name{econClusters}
\alias{econClusters}
\title{subsets by econ cluster}
\usage{
econClusters(arg)
}
\arguments{
\item{arg}{nil}
}
\description{
subsets by econ cluster
}
\examples{
econCluster()
}
\keyword{econ}
\keyword{level,}
\keyword{subset}
