% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installCensusFunctions.R
\name{installCensusFunctions}
\alias{installCensusFunctions}
\title{Helper function to change directories. v.1.2}
\usage{
installCensusFunctions(location)
}
\arguments{
\item{location}{Either local or github.}
}
\description{
This function installs the local or github version of Israel Census helper functions.
}
\examples{
installCensusFunctions("local")
installCensusFunctions("github")
}
\keyword{dir,}
\keyword{directory}
