% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{loadWealthEffectPackages}
\alias{loadWealthEffectPackages}
\title{Loads the required packages for the
Housing Wealth Effect Analysis.
Provides for simple dependency management to aid in reproducibility.}
\usage{
loadWealthEffectPackages(load = TRUE)
}
\arguments{
\item{load}{Defaults to TRUE.}
}
\description{
Loads the required packages for the
Housing Wealth Effect Analysis.
Provides for simple dependency management to aid in reproducibility.
}
\examples{
loadWealthEffectPackages()
}
\keyword{dependencies,}
\keyword{library,}
\keyword{packageManagement,}
\keyword{packages,}
\keyword{require,}
\keyword{requirements}

