% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFamiliesList.R
\name{makeFamiliesList}
\alias{makeFamiliesList}
\title{Make list of family, individual and exp data to workspace v: 0.3}
\usage{
makeFamiliesList(arg = TRUE)
}
\arguments{
\item{source}{Source for data, 'raw' imports from .por files. 'saved' files were previously parsed in R.}
}
\description{
This function imports save data.
}
\examples{
makeFamiliesList()
makeFamilies(TRUE)
}
\keyword{import,}
\keyword{saved}
