% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRegions.R
\name{makeRegions}
\alias{makeRegions}
\title{Add a region column based on $CODELOC (city) in the original data.
v. 0.35
This function adds columns $regionOne and $regionTwo
regionOne puts 13 cities in Gush Dan.
regionTwo puts  2 cities in Gush Dan.}
\usage{
makeRegions(arg = TRUE)
}
\arguments{
\item{arg}{The arg which you will perform the function on.}
}
\description{
Add a region column based on $CODELOC (city) in the original data.
v. 0.35
This function adds columns $regionOne and $regionTwo
regionOne puts 13 cities in Gush Dan.
regionTwo puts  2 cities in Gush Dan.
}
\examples{
makeRegions()
makeRegions("data")
makeRegions("renters")
makeRegions("owners")
makeRegions("writeOut")
makeRegions("load")
}
\keyword{city,}
\keyword{region}
