% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxReproducible.R
\name{maxReproducible}
\alias{maxReproducible}
\title{Work with saved data or from raw census downloads? Defaults to saved. 1 for raw. 2 for saved. Takes one argument.}
\usage{
maxReproducible(level = 2)
}
\description{
Work with saved data or from raw census downloads? Defaults to saved. 1 for raw. 2 for saved. Takes one argument.
}
\examples{
importData(1)
importData(2)
importData()
}
\keyword{import,}
\keyword{reproducibility}
\keyword{saved,}
