% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxReproducible.R
\name{maxReproducible}
\alias{maxReproducible}
\title{Work with saved data or from raw census downloads?  v: 1.6
Defaults to saved. 1 for raw. 2 for saved. 3 to import more saved.  4. to write out what was made by two.}
\usage{
maxReproducible(level = 5)
}
\arguments{
\item{level}{1..6, defaults to 5}
}
\description{
Work with saved data or from raw census downloads?  v: 1.6
Defaults to saved. 1 for raw. 2 for saved. 3 to import more saved.  4. to write out what was made by two.
}
\examples{
maxReproducible(1)
maxReproducible(2)
maxReproducible(3)
maxReproducible(4)
maxReproducible(5)
maxReproducible(6)
maxReproducible()
}
\seealso{
Other setup: \code{\link{installCensusFunctions}},
  \code{\link{install_if_needed}}, \code{\link{preSetUp}},
  \code{\link{setUp}}
}
\keyword{import,}
\keyword{reproducibility}
\keyword{saved,}
