% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxReproducible.R
\name{maxReproducible}
\alias{maxReproducible}
\title{Work with saved data or from raw census downloads?  v: 0.5
Defaults to saved. 1 for raw. 2 for saved. 3 to create saved(2) from raw(1) Takes one argument.}
\usage{
maxReproducible(level = 2)
}
\description{
Work with saved data or from raw census downloads?  v: 0.5
Defaults to saved. 1 for raw. 2 for saved. 3 to create saved(2) from raw(1) Takes one argument.
}
\examples{
maxReproducible(1)
maxReproducible(2)
maxReproducible(3)
maxReproducible()
}
\keyword{import,}
\keyword{reproducibility}
\keyword{saved,}
