% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingAverage.R
\name{movingAverage}
\alias{movingAverage}
\title{Add Moving Average Column to YYYY. v. 1.4}
\usage{
movingAverage(arg = "load")
}
\arguments{
\item{arg}{The year that data will be returned for.}
}
\description{
This function creates a lagged homeP column.
}
\examples{
movingAverage()
requires:
1: censusFunctions::importData("versionF")
2: load("../savedData/homeprices.Rda")
3: source("../includes/regionalPrices.R")
movingAverage("load")
adds the tminus1 and tminus2 columns for homeP
movingAverage("whshock")
calculates the error or unexpected shock difference
between (homeP - tminus1) and (tminus1 - tminus2)
}
\seealso{
Other mutate: \code{\link{regionalSizePrices}}
}
\keyword{average,}
\keyword{houseP}
