% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownershipByCity.R
\name{ownershipByCity}
\alias{ownershipByCity}
\title{split ownersYYYY and rentersYYYY by city  v.2.2
This function creates subsets of renters and owners by city}
\usage{
ownershipByCity(arg = TRUE)
}
\arguments{
\item{arg}{Defaults to true.}
}
\description{
split ownersYYYY and rentersYYYY by city  v.2.2
This function creates subsets of renters and owners by city
}
\examples{
ownershipByCity()
ownershipByCity(TRUE)
ownershipByCity(FALSE)
}
\keyword{owners,}
\keyword{renters,}
\keyword{subset}
