% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preSetUp.R
\name{preSetUp}
\alias{preSetUp}
\title{initial setup code ; v.0.5 load frames with regions
This function wraps setup code}
\usage{
preSetUp(arg = TRUE)
}
\arguments{
\item{arg}{Defaults to true.}
}
\description{
initial setup code ; v.0.5 load frames with regions
This function wraps setup code
}
\examples{
preSetUp(TRUE)
preSetUp()
}
\seealso{
Other setup: \code{\link{installCensusFunctions}},
  \code{\link{install_if_needed}},
  \code{\link{maxReproducible}}, \code{\link{setUp}}
}
\keyword{setup}
