% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalSizePrices.R
\name{regionalSizePrices}
\alias{regionalSizePrices}
\title{Add houseP column based on fam:(yr, size, region) v. 1.7}
\usage{
regionalSizePrices(frame = TRUE)
}
\arguments{
\item{frame}{The frame: data/owners/renters}
}
\description{
This function adds $houseP column.
}
\examples{
regionalSizePrices()
regionalSizePrices("famYYYY")
regionalSizePrices("data")
regionalSizePrices("owners")
regionalSizePrices("renters")
}
\keyword{case}
\keyword{price,}
\keyword{regional,}
