% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalSizePrices.R
\name{regionalSizePrices}
\alias{regionalSizePrices}
\title{Add houseP column based on fam:(yr, size, region) v. 3.5
famYYYYb is latest version.
This function adds $houseP and $housePTwo columns.}
\usage{
regionalSizePrices(frame = TRUE)
}
\arguments{
\item{frame}{The frame: data/owners/renters}
}
\description{
Add houseP column based on fam:(yr, size, region) v. 3.5
famYYYYb is latest version.
This function adds $houseP and $housePTwo columns.
}
\examples{
regionalSizePrices()
regionalSizePrices("data")
regionalSizePrices("famYYYY")
regionalSizePrices("famYYYYb")
regionalSizePrices("owners")
regionalSizePrices("renters")
}
\seealso{
Other mutate: \code{\link{movingAverage}}
}
\keyword{case}
\keyword{price,}
\keyword{regional,}
