% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFunction.R
\name{sampleFunction}
\alias{sampleFunction}
\title{A sample function}
\usage{
sampleFunction(arg1)
}
\arguments{
\item{arg1}{"options are yes and no, enclosed in parens."}
}
\description{
This function is a template for structuring my code.
}
\examples{
sampleFunction("Y")
sampleFunction("N")
}
\keyword{generic,}
\keyword{sample,}
\keyword{template}

