% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R, R/sampleFunction.R
\name{sampleFunction}
\alias{sampleFunction}
\title{Import data to workspace}
\usage{
sampleFunction(arg1)

sampleFunction(arg1)
}
\arguments{
\item{arg1}{"options are yes and no, enclosed in parens."}

\item{source}{Source for data, "raw" imports from .por files. "saved" files were previously parsed in R.}
}
\description{
This function imports Israeli census data.

This function is a template for structuring my code.
}
\examples{
importData("raw")
importData("saved")
sampleFunction("Y")
sampleFunction("N")
}
\keyword{generic,}
\keyword{import,}
\keyword{por,}
\keyword{raw}
\keyword{sample,}
\keyword{saved,}
\keyword{spss,}
\keyword{template}

