% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{setUp}
\alias{setUp}
\title{Secondary setup code, replaces setupV7
This function wraps setup code}
\usage{
setUp(arg = 4)
}
\arguments{
\item{arg}{Defaults to 4.}
}
\description{
Secondary setup code, replaces setupV7
This function wraps setup code
}
\examples{
setUp()
setUp(1)
setUp(2)
setUp(3)
setUp(4)
}
\seealso{
Other setup: \code{\link{installCensusFunctions}},
  \code{\link{install_if_needed}}, \code{\link{preSetUp}}
}
\keyword{setup}
