% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockData.R
\name{stockData}
\alias{stockData}
\title{Import stock data from exchange}
\usage{
stockData(arg)
}
\arguments{
\item{arg}{import- get it, add- add it as a column to the data frame, write- write the new dataframe}
}
\description{
This function imports mid-year prices from the tlv-125 index.
}
\keyword{import,}
\keyword{non-housing}
\keyword{stocks,}
\keyword{tlv-125,}
\keyword{wealth,}
\keyword{write}
