% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByOwnership.R
\name{subsetByOwnership}
\alias{subsetByOwnership}
\title{Creates dataframes expYYYYowners and expYYYYrenters,
calls weighHouseholds() ;   v. 3.65
This function creates subsets of renters and owners. and key money}
\usage{
subsetByOwnership(arg = TRUE)
}
\arguments{
\item{arg}{Defaults to true.}
}
\description{
Creates dataframes expYYYYowners and expYYYYrenters,
calls weighHouseholds() ;   v. 3.65
This function creates subsets of renters and owners. and key money
}
\examples{
subsetByOwnership('listKeyMoney')
subsetByOwnership()
## this should be called after loadData("versionD")
subsetByOwnership('versionD')
## this was the previous version
subsetByOwnership('better')

subsetByOwnership('listOwners')
subsetByOwnership('listNotKey')
subsetByOwnership('listHouseholds')
subsetByOwnership('listRenters')
subsetByOwnership('listKeyMoney')
subsetByOwnership(1)
subsetByOwnership(FALSE)
}
\keyword{owners,}
\keyword{renters,}
\keyword{subset}
