% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetByOwnership.R
\name{subsetByOwnership}
\alias{subsetByOwnership}
\title{Creates dataframes expYYYYowners and expYYYYrenters,
calls weighHouseholds() ; 
This function creates subsets of renters and owners. and key money}
\usage{
subsetByOwnership(arg = TRUE)
}
\arguments{
\item{arg}{Defaults to true.}
}
\description{
Creates dataframes expYYYYowners and expYYYYrenters,
calls weighHouseholds() ; 
This function creates subsets of renters and owners. and key money
}
\examples{
subsetByOwnership()
subsetByOwnership(1)
subsetByOwnership(FALSE)
}
\keyword{owners,}
\keyword{renters,}
\keyword{subset}
