% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCities.R
\name{subsetCities}
\alias{subsetCities}
\title{subsets the yearly surveys by city of residence, v 3.4
defaults to arg = "cities"
replace plyr::subset() with dplyr::filter()}
\usage{
subsetCities(arg = "saved")
}
\arguments{
\item{arg}{nil}
}
\description{
subsets the yearly surveys by city of residence, v 3.4
defaults to arg = "cities"
replace plyr::subset() with dplyr::filter()
}
\examples{
subsetCities()
subsetCities("writeOut")
subsetCities("make1")
subsetCities("make2")
subsetCities("make3")
subsetCities("addregion")
subsetCities("writeOutRegions")
subsetCities("saved")
}
\seealso{
Other manipulate: \code{\link{calcNHC}},
  \code{\link{makeRegions}}, \code{\link{stockData}}
}
\keyword{city,}
\keyword{column}
\keyword{region,}
\keyword{size,}
\keyword{subset,}
