% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCities.R
\name{subsetCities}
\alias{subsetCities}
\title{subsets the yearly surveys by city of residence, v 0.83}
\usage{
subsetCities(arg)
}
\arguments{
\item{arg}{nil}
}
\description{
subsets the yearly surveys by city of residence, v 0.83
}
\examples{
subsetCities()
}
\keyword{city,}
\keyword{subset}
