% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetCities.R
\name{subsetCities}
\alias{subsetCities}
\title{subsets the yearly surveys by city of residence, v 0.99
defaults to arg = "cities"
replace plyr::subset() with dplyr::filter()}
\usage{
subsetCities(arg = "cities")
}
\arguments{
\item{arg}{nil}
}
\description{
subsets the yearly surveys by city of residence, v 0.99
defaults to arg = "cities"
replace plyr::subset() with dplyr::filter()
}
\examples{
subsetCities()
subsetCities("make")
subsetCities("writeOut")
subsetCities("saved")
}
\keyword{city,}
\keyword{subset}
