% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighHouseholds.R
\name{weighHouseholds}
\alias{weighHouseholds}
\title{Applies Weights to households.}
\usage{
weighHouseholds(arg1 = 1)
}
\arguments{
\item{arg1}{Defaults to TRUE.}
}
\description{
This function weighs Israeli census data.
}
\examples{
weighHouseholds(1)
weighHouseholds(0)
}
\keyword{aggregate,}
\keyword{data}
\keyword{process,}

